flatten repo to reduce complexity and reorganize codebase given shift from webhook/CD driven releaseforms to cmdline form generation @ releasetime
