--- Dockerfile.local
@@ -1,13 +1,0 @@
-# build stage
-FROM golang:1.13-alpine AS builder
-#RUN apk --no-cache add git
-COPY . /go/src/github.com/paymentdata/releaseforms
-WORKDIR /go/src/github.com/paymentdata/releaseforms
-RUN CGO_ENABLED=0 go build -ldflags="-s -w" .
-RUN mkdir pdfs
-
-# final stage
-FROM scratch
-COPY --from=builder /go/src/github.com/paymentdata/releaseforms/releaseforms /go/releaseforms
-
-ENTRYPOINT ["/go/releaseforms"]
