+++ README.md
@@ -1,53 +1,49 @@
-# objective:
+`releaseforms` **shall** implement a pipeline of specially trained gophers who themselves can construct associated documentation, providing stakeholders and management with release documentation that satisfied control-objective 5 requirements.
 
-to implement a service which subscribes to **release events** and publishes associated **release documents** which are structured to comply with [PCIDSS](https://pcicompliance.stanford.edu/sites/g/files/sbiybj7706/f/16._change_control_policy_0.pdf)
-
-# releaseforms
-
-_problem statement:_ 
-
-generating a document on every release can be cumbersome for an organization any way you look at it.
-
-*especially* if the process is manual, and requires personnel.
-
------
-
-`releaseforms` **shall** implement a pipeline of specially trained gophers who themselves can construct associated documentation, 
-
-providing stakeholders and management with mostly filled out forms which only need to be reviewed and signed off.
-
-# current state of affairs/disclaimer things:
-
-_be warned, this is NOT generating compliant forms right now_
-
-_be warned, this MAY NEVER generate compliant forms_
+## current state of affairs/disclaimer things:
 
 *this project is very much in it's infancy, and therefore carries a risk of never delivering on the objective.*
-
-
-## currently depends on:
-
-- [athenapds](https://github.com/arachnys/athenapdf)
- (kudos to the folks working on arachnys projects!! 👏👏👏)
 
 -----
 
 now the fun stuff...
 
+commandline utility to generate a list of change items which represents changes from one past commit and one commit (either most recent or more recent)
+
+1. Δ/delta
+
+    - generating a list of merged PRs from last released commit to latest commit being released, depends on local/PWD being the git repo for the related release.
+
+2. context
+
+    - for each introduced Δ, generate a change item which collectively represent the given release changelog.
+
+
+# Usage example:
+
+`./delta $lastDeploymentSHA | ./contextaggregator`, where lastDeploymentSHA is set to the prior releases Commit.
+
 ## how to use: [w/ provided defaults and minimal setup]
 
-#### Configure GitHub repo/org webhook:
+### run this once for setup:
+``` bash
+git clone https://github.com/paymentdata/releaseforms
+cd releaseforms/
+echo -e "REPO=releaseforms\nORG=paymentdata" > .env
+go test ./...
+```
 
-1 navigate to _Settings > Webhooks_ on GitHub
+### for each release: (where the previously released commit for the .git is set as $lastDeploymentSHA)
+``` bash
+lastDeploymentSHA=d95539a
+./delta $lastDeploymentSHA | ./contextaggregator
+```
 
-2 select **add webhook**
+### PDF Generation currently depends on:
 
-3 **PayloadURL**: `yourhost:8081/webhooks1`, **Content-Type**: `application/json`, **Secret**: `MyGitHubSuperSecretSecrect...?`
+- [athenapds](https://github.com/arachnys/athenapdf)
+ (kudos to the folks working on arachnys projects!! 👏👏👏)
 
-4 `go get github.com/paymentdata/releaseforms` and run w/ the sample `.env` in the ${PWD}
+run via
+**needs athenapdf instance running**: `docker run -p 8080:8080 --rm arachnysdocker/athenapdf-service`
 
-5 `./releaseforms`
-
-6 trigger webhook
-
-7 iterate as needed
