+++ cmd/contextaggregator/people.go
@@ -1,4 +1,4 @@
-package people
+package main
 
 import (
 	"encoding/json"
