+++ go.mod
@@ -5,7 +5,5 @@ go 1.13
 require (
 	github.com/google/go-github/v28 v28.1.1
 	github.com/joho/godotenv v1.3.0
-	github.com/stretchr/testify v1.4.0 // indirect
 	golang.org/x/oauth2 v0.0.0-20191122200657-5d9234df094c
-	gopkg.in/go-playground/webhooks.v5 v5.13.0
 )
